SNOW-925280 Drop go 1.19 support, add 1.21 support
