+++ .github/workflows/build-test.yml
@@ -2,7 +2,7 @@ name: Build and Test
 
 on:
     push:
-        braches:
+        branches:
             - master
         tags:
             - v*
@@ -32,7 +32,7 @@ jobs:
             fail-fast: false
             matrix:
                 cloud: [ 'AWS', 'AZURE', 'GCP' ]
-                go: [ '1.20', '1.19' ]
+                go: [ '1.20', '1.21' ]
         name: ${{ matrix.cloud }} Go ${{ matrix.go }} on Ubuntu
         steps:
             - uses: actions/checkout@v1
@@ -60,7 +60,7 @@ jobs:
             fail-fast: false
             matrix:
                 cloud: [ 'AWS', 'AZURE', 'GCP' ]
-                go: [ '1.20', '1.19' ]
+                go: [ '1.20', '1.21' ]
         name: ${{ matrix.cloud }} Go ${{ matrix.go }} on Mac
         steps:
             - name: Sync clock
@@ -89,7 +89,7 @@ jobs:
             fail-fast: false
             matrix:
                 cloud: [ 'AWS', 'AZURE', 'GCP' ]
-                go: [ '1.20', '1.19' ]
+                go: [ '1.20', '1.21' ]
         name: ${{ matrix.cloud }} Go ${{ matrix.go }} on Windows
         steps:
             - uses: actions/checkout@v1
